backend_django/requirements.txt
asgiref==3.10.0
dataclasses==0.6
Django==4.1.13
django-cors-headers==4.9.0
djangorestframework==3.16.1
djongo==1.3.6
dnspython==2.7.0
pymongo==4.15.3
pytz==2025.2
sqlparse==0.2.4
typing_extensions==4.15.0
